move gnome ext code, avoid subproccess
